--- components/nofrendo-esp32/gpiocontroller.h
@@ -1,7 +1,0 @@
-#ifndef GAMEPAD_H
-#define GAMEPAD_H
-
-int gpioReadInput();
-void gpiocontrollerInit();
-
-#endif
