--- components/nofrendo-esp32/psxcontroller.h
@@ -1,7 +1,0 @@
-#ifndef PSXCONTROLLER_H
-#define PSXCONTROLLER_H
-
-int psxReadInput();
-void psxcontrollerInit();
-
-#endif
