+++ main/controller.h
@@ -1,0 +1,7 @@
+#ifndef _CONTROLLER_H_
+#define _CONTROLLER_H_
+
+int controller_read_input();
+void controller_init();
+
+#endif // _CONTROLLER_H_
